requirements.txt
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.11.2
cloudinary==1.8.0
dj-database-url==0.4.2
Django==1.11.3
django-cors-headers==2.0.2
django-filter==1.1.0
djangorestframework==3.7.1
gunicorn==19.7.1
Markdown==2.6.9
mock==2.0.0
pbr==3.1.1
psycopg2==2.7.1
pycparser==2.18
pytz==2017.2
six==1.11.0
urllib3==1.22
whitenoise==3.3.0


